ep3/API/requirements.txt
flask==1.0.2
psycopg2-binary==2.8.3
